requirements.txt
flask==3.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
boto3==1.34.7
amazon-transcribe==0.5.0
eventlet==0.33.3
gunicorn==21.2.0 